requirements.txt
streamlit
pandas
mysql-connector-python
sqlalchemy
pymysql